more refactor updates
