+++ src/abstracts/resource-checks.ts
